added google-auth-strategy

+++ controllers/api/v1/posts_api.js
@@ -18,29 +18,20 @@ module.exports.destroy=async function(req,res){
       let post=  await Post.findByIdAndDelete(req.params.id);
           // autherization to delete post
           // .id means converting objectid into string
-        //   if(post.user==req.user.id){
+          if(post.user==req.user.id){
               // delete post
                post.remove();
                // delete comments of post
               await Comment.deleteMany({post:req.params.id});
-            //   if(req.xhr){
-                //   return res.status(200).json({
-                //       data:{
-                //           post_id:req.params.id
-                //       },
-                //       message:"Post Deleted!"
-                //   });
-            //   }
-            //   req.flash('success','Post and associated comment deleted');
-               
+            
               return res.json(200,{message:"post and associated deleted successfully"
               });
-        //   }
-//           else{
-//   req.flash('error','You cannot delete this post');
-  
-//               return res.redirect('/');
-//           }
+          }
+          else{
+              return res.json(401,{
+                message:'you cannot delete post!'
+              });
+          }
       }catch(err){
 //   req.flash('error',err);
